jszakmeister: tweak paren highlighting
Turn on show paren mode, to help highlight a paren when you're close to
it.  Also use the highlight-parentheses package to help show which set
of parens you're in.  Note: rainbow parens overrides this when in
clojure and emacs-lisp mode.  But it's very useful to have this kind of
highlighting in general.  So enable it everywhere.